Refactor `api-docs-builder` types to avoid circular imports
