+++ packages/api-docs-builder/ProjectSettings.ts
@@ -1,10 +1,8 @@
 import { ComponentClassDefinition } from '@mui/internal-docs-utils';
-import { ComponentInfo, HookInfo } from './buildApiUtils';
 import { CreateTypeScriptProjectOptions } from './utils/createTypeScriptProject';
 import { CreateDescribeablePropSettings } from './utils/createDescribeableProp';
-import { ReactApi as ComponentReactApi } from './ApiBuilders/ComponentApiBuilder';
-import { ReactApi as HookReactApi } from './ApiBuilders/HookApiBuilder';
-import { Slot } from './utils/parseSlotsAndClasses';
+import { ComponentReactApi, HookReactApi } from './types/ApiBuilder.types';
+import { Slot, ComponentInfo, HookInfo } from './types/utils.types';
 
 export type SortingStrategiesType = {
   /**
